ImageClassification-MobileNet/requirements.txt
numpy==2.2.2
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
tensorflow==2.18.0
